verkle: reorganize vekrle tests and add getStem tests
